Folded edge receives into one sample
