--- azure-iot-device/samples/async-edge-scenarios/receive_twin_desired_properties_patch.py
@@ -1,52 +1,0 @@
-# -------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License. See License.txt in the project root for
-# license information.
-# --------------------------------------------------------------------------
-
-import asyncio
-from six.moves import input
-from azure.iot.device.aio import IoTHubModuleClient
-
-
-async def main():
-    # The client object is used to interact with your Azure IoT hub.
-    module_client = IoTHubModuleClient.create_from_edge_environment()
-
-    # connect the client.
-    await module_client.connect()
-
-    # Define behavior for receiving a twin patch
-    # NOTE: this could be a function or a coroutine
-    def twin_patch_handler(patch):
-        print("the data in the desired properties patch was: {}".format(patch))
-
-    # set the twin patch handler on the client
-    module_client.on_twin_desired_properties_patch_received = twin_patch_handler
-
-    # define behavior for halting the application
-    def stdin_listener():
-        while True:
-            selection = input("Press Q to quit\n")
-            if selection == "Q" or selection == "q":
-                print("Quitting...")
-                break
-
-    # Run the stdin listener in the event loop
-    loop = asyncio.get_running_loop()
-    user_finished = loop.run_in_executor(None, stdin_listener)
-
-    # Wait for user to indicate they are done listening for messages
-    await user_finished
-
-    # Finally, shut down the client
-    await module_client.shutdown()
-
-
-if __name__ == "__main__":
-    asyncio.run(main())
-
-    # If using Python 3.6 or below, use the following code instead of asyncio.run(main()):
-    # loop = asyncio.get_event_loop()
-    # loop.run_until_complete(main())
-    # loop.close()
